+++ src/biodwh2-sider/src/main/java/de/unibi/agbi/biodwh2/sider/etl/SiderGraphExporter.java
@@ -12,6 +12,7 @@ import de.unibi.agbi.biodwh2.core.model.graph.IndexDescription;
 import de.unibi.agbi.biodwh2.core.model.graph.Node;
 import de.unibi.agbi.biodwh2.sider.SiderDataSource;
 import de.unibi.agbi.biodwh2.sider.model.*;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -28,6 +29,7 @@ import java.util.*;
  */
 public class SiderGraphExporter extends GraphExporter<SiderDataSource> {
     private static final Logger LOGGER = LoggerFactory.getLogger(SiderGraphExporter.class);
+    private static final String DRUG_LABEL = "Drug";
     private static final String MEDDRA_TERM_LABEL = "MeddraTerm";
     private static final String UMLS_ID_KEY = "umls_id";
     private static final String MEDDRA_ID_KEY = "meddra_id";
@@ -35,50 +37,47 @@ public class SiderGraphExporter extends GraphExporter<SiderDataSource> {
     private static final String STEREO_ID_KEY = "stereo_id";
     private static final String ATC_CODES_KEY = "atc_codes";
 
-    private final Set<Long> edgeHashCache = new HashSet<>();
-
     public SiderGraphExporter(final SiderDataSource dataSource) {
         super(dataSource);
     }
 
     @Override
     public long getExportVersion() {
-        return 1;
+        return 2;
     }
 
     @Override
     protected boolean exportGraph(final Workspace workspace, final Graph graph) throws ExporterException {
-        graph.addIndex(IndexDescription.forNode("Drug", FLAT_ID_KEY, IndexDescription.Type.NON_UNIQUE));
-        graph.addIndex(IndexDescription.forNode("Drug", STEREO_ID_KEY, IndexDescription.Type.NON_UNIQUE));
+        graph.addIndex(IndexDescription.forNode(DRUG_LABEL, FLAT_ID_KEY, IndexDescription.Type.NON_UNIQUE));
+        graph.addIndex(IndexDescription.forNode(DRUG_LABEL, STEREO_ID_KEY, IndexDescription.Type.UNIQUE));
         graph.addIndex(IndexDescription.forNode(MEDDRA_TERM_LABEL, UMLS_ID_KEY, IndexDescription.Type.NON_UNIQUE));
-        graph.addIndex(IndexDescription.forNode(MEDDRA_TERM_LABEL, MEDDRA_ID_KEY, IndexDescription.Type.NON_UNIQUE));
+        graph.addIndex(IndexDescription.forNode(MEDDRA_TERM_LABEL, MEDDRA_ID_KEY, IndexDescription.Type.UNIQUE));
+        final Map<String, CombinedDrug> flatIdDrugMap = new HashMap<>();
         if (LOGGER.isInfoEnabled())
-            LOGGER.info("Exporting drug names...");
-        final Map<String, Long> drugIdNodeIdMap = addAllDrugNames(workspace, graph);
+            LOGGER.info("Collecting drug atc codes...");
+        populateCombinedDrugsWithATC(workspace, flatIdDrugMap);
         if (LOGGER.isInfoEnabled())
-            LOGGER.info("Exporting drug atc codes...");
-        addAllDrugAtcCodes(workspace, graph, drugIdNodeIdMap);
+            LOGGER.info("Collecting drug names...");
+        populateCombinedDrugsWithName(workspace, flatIdDrugMap);
         if (LOGGER.isInfoEnabled())
             LOGGER.info("Exporting indications...");
-        final Map<String, Long> meddraIdNodeIdMap = new HashMap<>();
-        addAllIndications(workspace, graph, drugIdNodeIdMap, meddraIdNodeIdMap);
+        addAllIndications(workspace, graph, flatIdDrugMap);
         if (LOGGER.isInfoEnabled())
             LOGGER.info("Exporting side effects...");
-        addAllSideEffects(workspace, graph, drugIdNodeIdMap, meddraIdNodeIdMap);
+        addAllSideEffects(workspace, graph, flatIdDrugMap);
         if (LOGGER.isInfoEnabled())
             LOGGER.info("Exporting side effect frequencies...");
-        addAllSideEffectFrequencies(workspace, graph, drugIdNodeIdMap, meddraIdNodeIdMap);
-        edgeHashCache.clear();
+        addAllSideEffectFrequencies(workspace, graph, flatIdDrugMap);
         return true;
     }
 
-    private Map<String, Long> addAllDrugNames(final Workspace workspace, final Graph graph) throws ExporterException {
-        final Map<String, Long> drugIdNodeIdMap = new HashMap<>();
-        for (final DrugName drug : parseTsvFile(workspace, SiderUpdater.DRUG_NAMES_FILE_NAME, DrugName.class)) {
-            final Node node = graph.addNodeFromModel(drug);
-            drugIdNodeIdMap.put(drug.id, node.getId());
+    private void populateCombinedDrugsWithATC(final Workspace workspace,
+                                              final Map<String, CombinedDrug> flatIdDrugMap) {
+        for (final DrugAtc atcCode : parseTsvFile(workspace, SiderUpdater.DRUG_ATC_FILE_NAME, DrugAtc.class)) {
+            if (!flatIdDrugMap.containsKey(atcCode.id))
+                flatIdDrugMap.put(atcCode.id, new CombinedDrug(atcCode.id));
+            flatIdDrugMap.get(atcCode.id).atcCodes.add(atcCode.atc);
         }
-        return drugIdNodeIdMap;
     }
 
     private <T> Iterable<T> parseTsvFile(final Workspace workspace, final String fileName,
@@ -91,77 +90,28 @@ public class SiderGraphExporter extends GraphExporter<SiderDataSource> {
         }
     }
 
-    private void addAllDrugAtcCodes(final Workspace workspace, final Graph graph,
-                                    final Map<String, Long> drugIdNodeIdMap) throws ExporterException {
-        for (final DrugAtc atcCode : parseTsvFile(workspace, SiderUpdater.DRUG_ATC_FILE_NAME, DrugAtc.class)) {
-            final Node drugNode = graph.getNode(drugIdNodeIdMap.get(atcCode.id));
-            addValueToNodePropertySet(drugNode, ATC_CODES_KEY, atcCode.atc);
-            graph.update(drugNode);
+    private void populateCombinedDrugsWithName(final Workspace workspace,
+                                               final Map<String, CombinedDrug> flatIdDrugMap) {
+        for (final DrugName drug : parseTsvFile(workspace, SiderUpdater.DRUG_NAMES_FILE_NAME, DrugName.class)) {
+            if (!flatIdDrugMap.containsKey(drug.id))
+                flatIdDrugMap.put(drug.id, new CombinedDrug(drug.id));
+            flatIdDrugMap.get(drug.id).name = drug.name;
         }
-    }
-
-    private static void addValueToNodePropertySet(final Node node, final String key, final String value) {
-        final Set<String> values = new HashSet<>();
-        if (node.hasProperty(key))
-            Collections.addAll(values, node.getProperty(key));
-        values.add(value);
-        node.setProperty(key, values.toArray(new String[0]));
     }
 
     private void addAllIndications(final Workspace workspace, final Graph graph,
-                                   final Map<String, Long> drugIdNodeIdMap,
-                                   final Map<String, Long> meddraIdNodeIdMap) throws ExporterException {
-        edgeHashCache.clear();
-        for (Indication indication : parseGzipTsvFile(workspace, SiderUpdater.INDICATIONS_FILE_NAME,
-                                                      Indication.class)) {
-            long meddraTermNodeId = getOrAddMeddraTermNode(graph, meddraIdNodeIdMap, indication.umlsConceptId,
-                                                           indication.conceptName, indication.meddraUmlsConceptId,
-                                                           indication.meddraConceptName);
-            long drugNodeId = getOrAddDrugNode(graph, indication.flatCompoundId, indication.stereoCompoundId,
-                                               drugIdNodeIdMap);
-            long hash = Objects.hash(meddraTermNodeId, drugNodeId, indication.label, indication.detectionMethod);
-            if (!edgeHashCache.contains(hash)) {
-                graph.addEdge(drugNodeId, meddraTermNodeId, "INDICATES", "label", indication.label, "detection_method",
-                              indication.detectionMethod);
-                edgeHashCache.add(hash);
-            }
+                                   final Map<String, CombinedDrug> flatIdDrugMap) throws ExporterException {
+        for (final Indication indication : parseGzipTsvFile(workspace, SiderUpdater.INDICATIONS_FILE_NAME,
+                                                            Indication.class)) {
+            if (skipConcept(indication.meddraConceptType))
+                continue;
+            final Node drugNode = getOrAddDrugNode(graph, flatIdDrugMap, indication.flatCompoundId,
+                                                   indication.stereoCompoundId);
+            final Node termNode = getOrAddTermNode(graph, indication.umlsConceptId, indication.conceptName,
+                                                   indication.meddraUmlsConceptId, indication.meddraConceptName);
+            graph.addEdge(drugNode, termNode, "INDICATES", "label", indication.label, "detection_method",
+                          indication.detectionMethod);
         }
-    }
-
-    private long getOrAddMeddraTermNode(final Graph graph, final Map<String, Long> meddraIdNodeIdMap,
-                                        final String umlsConceptId, final String umlsConceptName,
-                                        final String conceptId, final String conceptName) {
-        if (meddraIdNodeIdMap.containsKey(conceptId))
-            return meddraIdNodeIdMap.get(conceptId);
-        Node node;
-        if (umlsConceptName == null)
-            node = graph.addNode(MEDDRA_TERM_LABEL, MEDDRA_ID_KEY, conceptId, UMLS_ID_KEY, umlsConceptId, "meddra_name",
-                                 conceptName);
-        else
-            node = graph.addNode(MEDDRA_TERM_LABEL, MEDDRA_ID_KEY, conceptId, UMLS_ID_KEY, umlsConceptId, "meddra_name",
-                                 conceptName, "umls_name", umlsConceptName);
-        meddraIdNodeIdMap.put(conceptId, node.getId());
-        return node.getId();
-    }
-
-    private long getOrAddDrugNode(final Graph graph, final String flatId, final String stereoId,
-                                  final Map<String, Long> drugIdNodeIdMap) {
-        Long nodeId = drugIdNodeIdMap.get(flatId);
-        if (nodeId == null)
-            nodeId = drugIdNodeIdMap.get(stereoId);
-        if (nodeId != null) {
-            final Node node = graph.getNode(nodeId);
-            if (!node.hasProperty(STEREO_ID_KEY)) {
-                node.setProperty(STEREO_ID_KEY, stereoId);
-                graph.update(node);
-                drugIdNodeIdMap.put(stereoId, node.getId());
-            }
-            return nodeId;
-        }
-        final Node node = graph.addNode("Drug", FLAT_ID_KEY, flatId, STEREO_ID_KEY, stereoId);
-        drugIdNodeIdMap.put(flatId, node.getId());
-        drugIdNodeIdMap.put(stereoId, node.getId());
-        return node.getId();
     }
 
     private <T> Iterable<T> parseGzipTsvFile(final Workspace workspace, final String fileName,
@@ -174,48 +124,82 @@ public class SiderGraphExporter extends GraphExporter<SiderDataSource> {
         }
     }
 
-    private void addAllSideEffects(final Workspace workspace, final Graph graph,
-                                   final Map<String, Long> drugIdNodeIdMap,
-                                   final Map<String, Long> meddraIdNodeIdMap) throws ExporterException {
-        edgeHashCache.clear();
-        for (SideEffect sideEffect : parseGzipTsvFile(workspace, SiderUpdater.SIDE_EFFECTS_FILE_NAME,
-                                                      SideEffect.class)) {
-            long meddraTermNodeId = getOrAddMeddraTermNode(graph, meddraIdNodeIdMap, sideEffect.umlsConceptId, null,
-                                                           sideEffect.meddraUmlsConceptId, sideEffect.sideEffectName);
-            long drugNodeId = getOrAddDrugNode(graph, sideEffect.flatCompoundId, sideEffect.stereoCompoundId,
-                                               drugIdNodeIdMap);
-            long hash = Objects.hash(meddraTermNodeId, drugNodeId, sideEffect.label);
-            if (!edgeHashCache.contains(hash)) {
-                graph.addEdge(drugNodeId, meddraTermNodeId, "HAS_SIDE_EFFECT", "label", sideEffect.label);
-                edgeHashCache.add(hash);
+    private boolean skipConcept(final String conceptType) {
+        return StringUtils.isNotEmpty(conceptType) && !"PT".equals(conceptType);
+    }
+
+    private Node getOrAddDrugNode(final Graph graph, final Map<String, CombinedDrug> flatIdDrugMap, final String flatId,
+                                  final String stereoId) {
+        Node node = graph.findNode(DRUG_LABEL, STEREO_ID_KEY, stereoId);
+        if (node == null) {
+            final CombinedDrug drug = flatIdDrugMap.get(flatId);
+            if (drug != null) {
+                if (drug.name != null && drug.atcCodes.size() > 0)
+                    node = graph.addNode(DRUG_LABEL, STEREO_ID_KEY, stereoId, FLAT_ID_KEY, flatId, "name", drug.name,
+                                         ATC_CODES_KEY, drug.atcCodes.toArray(new String[0]));
+                else if (drug.name != null)
+                    node = graph.addNode(DRUG_LABEL, STEREO_ID_KEY, stereoId, FLAT_ID_KEY, flatId, "name", drug.name);
+                else if (drug.atcCodes.size() > 0)
+                    node = graph.addNode(DRUG_LABEL, STEREO_ID_KEY, stereoId, FLAT_ID_KEY, flatId, ATC_CODES_KEY,
+                                         drug.atcCodes.toArray(new String[0]));
+            } else
+                node = graph.addNode(DRUG_LABEL, STEREO_ID_KEY, stereoId, FLAT_ID_KEY, flatId);
+        }
+        return node;
+    }
+
+    private Node getOrAddTermNode(final Graph graph, final String umlsConceptId, final String umlsConceptName,
+                                  final String meddraUmlsConceptId, final String meddraConceptName) {
+        Node termNode = null;
+        if (StringUtils.isNotEmpty(meddraUmlsConceptId)) {
+            termNode = graph.findNode(MEDDRA_TERM_LABEL, MEDDRA_ID_KEY, meddraUmlsConceptId);
+            if (termNode == null) {
+                termNode = graph.addNode(MEDDRA_TERM_LABEL, MEDDRA_ID_KEY, meddraUmlsConceptId, UMLS_ID_KEY,
+                                         umlsConceptId, "meddra_name", meddraConceptName, "umls_name", umlsConceptName);
             }
+        } else {
+            final Iterable<Node> umlsTermNodes = graph.findNodes(MEDDRA_TERM_LABEL, UMLS_ID_KEY, umlsConceptId);
+            for (final Node candidate : umlsTermNodes)
+                if (candidate.getProperty(MEDDRA_ID_KEY) == null)
+                    termNode = candidate;
+            if (termNode == null)
+                termNode = graph.addNode(MEDDRA_TERM_LABEL, UMLS_ID_KEY, umlsConceptId, "umls_name", umlsConceptName);
+        }
+        return termNode;
+    }
+
+    private void addAllSideEffects(final Workspace workspace, final Graph graph,
+                                   final Map<String, CombinedDrug> flatIdDrugMap) throws ExporterException {
+        for (final SideEffect sideEffect : parseGzipTsvFile(workspace, SiderUpdater.SIDE_EFFECTS_FILE_NAME,
+                                                            SideEffect.class)) {
+            if (skipConcept(sideEffect.meddraConceptType))
+                continue;
+            final Node drugNode = getOrAddDrugNode(graph, flatIdDrugMap, sideEffect.flatCompoundId,
+                                                   sideEffect.stereoCompoundId);
+            final Node termNode = getOrAddTermNode(graph, sideEffect.umlsConceptId, null,
+                                                   sideEffect.meddraUmlsConceptId, sideEffect.sideEffectName);
+            graph.addEdge(drugNode, termNode, "HAS_SIDE_EFFECT", "label", sideEffect.label);
         }
     }
 
     private void addAllSideEffectFrequencies(final Workspace workspace, final Graph graph,
-                                             final Map<String, Long> drugIdNodeIdMap,
-                                             final Map<String, Long> meddraIdNodeIdMap) throws ExporterException {
-        edgeHashCache.clear();
-        for (Frequency frequency : parseGzipTsvFile(workspace, SiderUpdater.FREQUENCIES_FILE_NAME, Frequency.class)) {
-            long meddraTermNodeId = getOrAddMeddraTermNode(graph, meddraIdNodeIdMap, frequency.umlsConceptId, null,
-                                                           frequency.meddraUmlsConceptId, frequency.sideEffectName);
-            long drugNodeId = getOrAddDrugNode(graph, frequency.flatCompoundId, frequency.stereoCompoundId,
-                                               drugIdNodeIdMap);
-            long hash = Objects.hash(meddraTermNodeId, drugNodeId, frequency.frequency, frequency.frequencyLowerBound,
-                                     frequency.frequencyUpperBound, frequency.placebo);
-            if (!edgeHashCache.contains(hash)) {
-                EdgeBuilder builder = graph.buildEdge();
-                builder.fromNode(drugNodeId);
-                builder.toNode(meddraTermNodeId);
-                builder.withLabel("HAS_SIDE_EFFECT_FREQUENCY");
-                builder.withProperty("frequency", frequency.frequency);
-                builder.withProperty("frequency_lower_bound", frequency.frequencyLowerBound);
-                builder.withProperty("frequency_upper_bound", frequency.frequencyUpperBound);
-                if (frequency.placebo != null)
-                    builder.withProperty("placebo", frequency.placebo.equalsIgnoreCase("placebo"));
-                builder.build();
-                edgeHashCache.add(hash);
-            }
+                                             final Map<String, CombinedDrug> flatIdDrugMap) throws ExporterException {
+        for (final Frequency frequency : parseGzipTsvFile(workspace, SiderUpdater.FREQUENCIES_FILE_NAME,
+                                                          Frequency.class)) {
+            if (skipConcept(frequency.meddraConceptType))
+                continue;
+            final Node drugNode = getOrAddDrugNode(graph, flatIdDrugMap, frequency.flatCompoundId,
+                                                   frequency.stereoCompoundId);
+            final Node termNode = getOrAddTermNode(graph, frequency.umlsConceptId, null, frequency.meddraUmlsConceptId,
+                                                   frequency.sideEffectName);
+            EdgeBuilder builder = graph.buildEdge().fromNode(drugNode).toNode(termNode);
+            builder.withLabel("HAS_SIDE_EFFECT_FREQUENCY");
+            builder.withProperty("frequency", frequency.frequency);
+            builder.withProperty("frequency_lower_bound", frequency.frequencyLowerBound);
+            builder.withProperty("frequency_upper_bound", frequency.frequencyUpperBound);
+            if (frequency.placebo != null)
+                builder.withProperty("placebo", frequency.placebo.equalsIgnoreCase("placebo"));
+            builder.build();
         }
     }
 }